pyproject.toml
[project]
name = "repl-nix-smartscrapenlp"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "openai>=1.55.3",
    "psycopg2-binary>=2.9.10",
    "trafilatura>=1.12.2",
    "beautifulsoup4>=4.12.3",
    "scikit-learn>=1.5.2",
    "pillow>=11.0.0",
    "requests>=2.32.3",
]
